HLRC: split tasks request converters
In an effort to encapsulate the different clients, the request
converters are being shuffled around. This splits the TasksClient
request converters.
